7일 이전 실패 메세지 api 구현, sendMessage 구현 중

+++ app/src/main/java/com/coinlive/chat/firebase/model/Ama.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Ama(
     @SerializedName("pause") val isPause:Boolean,
-    @SerializedName("start") val startTime:Int,
-    @SerializedName("end") val endTime:Int?,
+    @SerializedName("start") val startTime:Long,
+    @SerializedName("end") val endTime:Long?,
     @SerializedName("heart") val heartCount:Int
 )
